+++ files/types.d.ts
@@ -242,10 +242,11 @@ declare namespace R {
       list: any[]
     ): boolean
 
-    isType(targetType: RambdaTypes, input: any): boolean
     isArray(input: any): boolean
-    isString(input: any): boolean
+    isFalsy(input: any): boolean
     isObject(input: any): boolean
+    isString(input: any): boolean
+    isType(targetType: RambdaTypes, input: any): boolean
 
     isPromise(
       maybePromiseOrAsync: any