+++ files/Changelog.md
@@ -2,6 +2,8 @@
 
 1.7.0 Rename `R.then` to `R.resolve` because of Ramda issue with `R.then`(they rename it to `R.andThen`) 
 
+- Add `R.isFalsy`, `R.nextIndex`
+
 1.6.3 Forgot to export `R.unless`(credit to @mobily for the PR)
 
 1.6.2 NPM for no reason doesn't respect the changed code and `R.compact` is missing