+++ files/Changelog.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
-1.7.0 Rename `R.then` to `R.resolve` because of Ramda issue with `R.then`(they rename it to `R.andThen`) 
+1.7.0 Rename `R.then` to `R.resolve` because of Ramda issue with `R.then`(they rename it to `R.andThen`)
 
-- Add `R.isFalsy`, `R.nextIndex`
+- Add `R.isFalsy`, `R.nextIndex` and `R.mergeDeep`
 
 1.6.3 Forgot to export `R.unless`(credit to @mobily for the PR)